requirements.txt
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langsmith==0.2.7
langchain-openai==0.2.14
langchain-google-genai==2.0.8
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.10.4
RapidFuzz==3.14.3
pytest==8.3.4
